Add files via upload
To use this Brain Tumor Segmentation project, first ensure that MATLAB is installed on your system with the Image Processing Toolbox. Download or clone the project files and open the folder in MATLAB. Place your MRI brain scan images in the MRI_Images folder. Then, open the main script file (Segmentation_Code.m) and run it. The code will automatically process the image by converting it to grayscale, filtering noise, segmenting the tumor region, and highlighting it. The final output will be displayed and saved in the Results folder. You can also customize the code to test different images or modify processing steps as needed.